requirements.txt
nose
pymatgen
pymatgen-db>=0.4.1
pymongo
matplotlib
shapely
